+++ sugarloaf/benches/bench_sugar_pile.rs
@@ -65,21 +65,21 @@ fn bench_sugar_pile(c: &mut Criterion) {
                 sugarloaf.insert_on_current_line(&Sugar {
                     content: ' ',
                     foreground_color: [0.0, 0.0, 0.0, 1.0],
-                    background_color: [0.0, 1.0, 1.0, 1.0],
+                    background_color: Some([0.0, 1.0, 1.0, 1.0]),
                     ..Sugar::default()
                 });
 
                 sugarloaf.insert_on_current_line(&Sugar {
                     content: '«',
                     foreground_color: [0.0, 0.0, 0.0, 1.0],
-                    background_color: [0.0, 1.0, 1.0, 1.0],
+                    background_color: Some([0.0, 1.0, 1.0, 1.0]),
                     ..Sugar::default()
                 });
 
                 sugarloaf.insert_on_current_line(&Sugar {
                     content: '≥',
                     foreground_color: [0.0, 0.0, 0.0, 1.0],
-                    background_color: [0.0, 1.0, 1.0, 1.0],
+                    background_color: Some([0.0, 1.0, 1.0, 1.0]),
                     ..Sugar::default()
                 });
             }