+++ sugarloaf/benches/bench_sugar_pile_with_screen.rs
@@ -78,7 +78,7 @@ fn bench_sugar_pile_with_screen(c: &mut Criterion) {
                                 sugarloaf.insert_on_current_line(&Sugar {
                                     content: 'a',
                                     foreground_color: [1.0, 1.0, 1.0, 1.0],
-                                    background_color: [0.0, 1.0, 1.0, 1.0],
+                                    background_color: Some([0.0, 1.0, 1.0, 1.0]),
                                     ..Sugar::default()
                                 });
                             }