+++ sugarloaf/examples/text.rs
@@ -66,43 +66,43 @@ async fn main() {
             Sugar {
                 content: 'S',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'u',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 1.0, 1.0],
+                background_color: Some([1.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'g',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'a',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 1.0, 1.0],
+                background_color: Some([1.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'r',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'g',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 0.0, 1.0, 1.0],
+                background_color: Some([0.0, 0.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: '|',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 1.0, 1.0],
+                background_color: Some([1.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
         ];
@@ -111,37 +111,37 @@ async fn main() {
             Sugar {
                 content: 'l',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'o',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 1.0, 1.0],
+                background_color: Some([1.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'a',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'f',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 0.0, 1.0, 1.0],
+                background_color: Some([0.0, 0.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'g',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 0.0, 1.0, 1.0],
+                background_color: Some([0.0, 0.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: '|',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 1.0, 1.0],
+                background_color: Some([1.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
         ];
@@ -150,46 +150,46 @@ async fn main() {
             Sugar {
                 content: ' ',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'r',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 0.0, 1.0, 1.0],
+                background_color: Some([0.0, 0.0, 1.0, 1.0]),
                 decoration: SugarDecoration::Underline,
                 ..Sugar::default()
             },
             Sugar {
                 content: 'i',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 decoration: SugarDecoration::Underline,
                 ..Sugar::default()
             },
             Sugar {
                 content: 'o',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 1.0, 1.0],
+                background_color: Some([1.0, 1.0, 1.0, 1.0]),
                 decoration: SugarDecoration::Underline,
                 ..Sugar::default()
             },
             Sugar {
                 content: 'g',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 1.0, 0.0, 1.0],
+                background_color: Some([0.0, 1.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: '¼',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 0.0, 1.0],
+                background_color: Some([1.0, 1.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: '¬',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 1.0, 0.0, 1.0],
+                background_color: Some([0.0, 1.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
         ];
@@ -199,34 +199,34 @@ async fn main() {
             Sugar {
                 content: '㏑',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 1.0, 1.0, 1.0],
+                background_color: Some([0.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             // Font Symbol (apple symbols font)
             Sugar {
                 content: '⫹',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             // Font Regular (firamono)
             Sugar {
                 content: 'λ',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 1.0, 1.0, 1.0],
+                background_color: Some([0.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             // Font Emojis
             Sugar {
                 content: '🥇',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: '👷',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 0.0, 1.0, 1.0],
+                background_color: Some([0.0, 0.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
         ];
@@ -236,31 +236,31 @@ async fn main() {
             Sugar {
                 content: 'a',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 1.0, 1.0],
+                background_color: Some([1.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: '％',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [0.0, 1.0, 1.0, 1.0],
+                background_color: Some([0.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: '',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.5, 0.5, 0.5, 1.0],
+                background_color: Some([0.5, 0.5, 0.5, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: 'a',
                 foreground_color: [0.0, 0.0, 0.0, 1.0],
-                background_color: [1.0, 1.0, 1.0, 1.0],
+                background_color: Some([1.0, 1.0, 1.0, 1.0]),
                 ..Sugar::default()
             },
             Sugar {
                 content: '',
                 foreground_color: [1.0, 1.0, 1.0, 1.0],
-                background_color: [0.0, 0.0, 0.0, 1.0],
+                background_color: Some([0.0, 0.0, 0.0, 1.0]),
                 ..Sugar::default()
             },
         ];